24-08-22

+++ home/views.py
@@ -1,8 +1,15 @@
 from itertools import count
 from django.shortcuts import render,redirect
-from .models import Room, Topic
+from django.http import HttpResponse
+from .models import Room, Topic, Message
 from .forms import RoomForm
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login,logout
+from django.contrib import messages
+from django.contrib.auth.forms import UserCreationForm
+
 
 # Create your views here.
 
@@ -14,6 +21,57 @@ from django.db.models import Q
 #     {'id':3,'name':'Irfan'},
 
 # ]
+
+
+def login_regis(request):
+    page='login'
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+        
+    if request.method=='POST':
+        username=request.POST.get('username').lower()
+        password=request.POST.get('password')
+        try:
+            user=User.objects.get(username=username)
+        except:
+            messages.error(request, 'User doesnt exit!!!')
+
+        user=authenticate(request,username=username,password=password)
+        if user!=None:
+            login(request,user)
+            return redirect('home')
+            
+        else:
+            messages.error(request, 'Wrong Password')
+
+    context={'page':page}
+    return render(request,'home/login_regis.html',context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def register(request):
+    page='register'
+    form=UserCreationForm()
+    if request.method=='POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Try again')
+            return redirect('home')
+
+
+    context={'page':page,'form':form}
+    return render(request,'home/login_regis.html',context)
+
 def home(request):
     topics=Topic.objects.all()
     rom=Room.objects.all()
@@ -46,12 +104,20 @@ def home(request):
 
 def room(request,pk):
     room=Room.objects.get(id=pk)
+    if request.method=="POST":
+        message=request.POST.get('message')
+        Message.objects.create(user=request.user,room=room,body=message)
+        return redirect('room',pk=room.id)
+    
 
-    context={'room':room}
+
+
+    messages=room.message_set.all().order_by('-created')
+    context={'room':room,'messages':messages}
     return render(request,'home/room.html',context)
 
 
-
+@login_required(login_url='login_regis')
 def createRoom(request):
     form=RoomForm()
     if request.method=="POST":
@@ -63,10 +129,14 @@ def createRoom(request):
     return render(request,'home/room_form.html',context)
 
 
-
+@login_required(login_url='login_regis')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse("Not allowed!")
+        
+
     if request.method=="POST":
         form=RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -81,10 +151,15 @@ def updateRoom(request,pk):
 #     room.delete()
 #     return redirect('home')
 
-
+@login_required(login_url='login_regis')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("Not allowed!")
     if request.method=='POST':
         room.delete()
         return redirect('home')
     return render(request,'home/delete.html',{'obj':room})
+
+
+
